refactor and more files around
